split product model
